% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excalibdata.R
\docType{data}
\name{excalibdata}
\alias{excalibdata}
\title{The dataset `excalibdata.Rdata` contains example data from an evaluation of
an assay measuring recency of infection. At an assay result of <10, the
specimen is considered to be recently infected. It further contains viral
load data, which is commonly used to reduce false recency. For example,
when recency is defined as assay result <10 and viral load > 1000, the FRR
is substantially lower (but the MDRI is also reduced).}
\format{An object of class \code{data.frame} with 1460 rows and 7 columns.}
\usage{
excalibdata
}
\description{
The dataset `excalibdata.Rdata` contains example data from an evaluation of
an assay measuring recency of infection. At an assay result of <10, the
specimen is considered to be recently infected. It further contains viral
load data, which is commonly used to reduce false recency. For example,
when recency is defined as assay result <10 and viral load > 1000, the FRR
is substantially lower (but the MDRI is also reduced).
}
\keyword{datasets}
