% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_mdri.R
\name{mdrical}
\alias{mdrical}
\title{Estimate MDRI (point estimate and confidence interval) using binomial regression and a maximum likelihood approach}
\usage{
mdrical(data = data, subid_var = "sid", time_var = "time",
  functional_forms = c("cloglog_linear", "logit_cubic"),
  recency_cutoff_time = 730.5, inclusion_time_threshold = 800,
  recency_rule = "binary_data", recency_vars = "recency_status",
  recency_params = NULL, n_bootstraps = 100, alpha = 0.05, plot = TRUE,
  parallel = FALSE, cores = 4)
}
\arguments{
\item{data}{A data frame containing variables for subject identifier, time (since detectable infection), and variables with biomarker readings or recency status (to be specified in recency_vars)}

\item{subid_var}{The variable in the dataframe identifying subjects}

\item{time_var}{The variable in the dataframe indicating time between 'time zero' (usually detectable infection) and biomarker measurement}

\item{functional_forms}{Select functional form/link function combinations for fitting probability of testing recent as a function of time to data using binomial regression
(see Details). Default=all supported functional forms.}

\item{recency_cutoff_time}{Recency time cut-off ('Big T'). Default=730.5.}

\item{inclusion_time_threshold}{Data points beyond this time are excluded from the calculation (in same unit as recency_cutoff_time, default=800).}

\item{recency_rule}{Specified rule for defining recent/non-recent outcomes from biomarker data (see Details)}

\item{recency_vars}{Variables to be used in determining recency outcomes}

\item{recency_params}{Vector of numeric parameters (e.g. thresholds) for determining recency according to the relevant rule}

\item{n_bootstraps}{Number of subject-level bootstrap resampling operations for estimating confidence intervals, default=100 (useful for testing purposes only)}

\item{alpha}{Confidence level, default=0.05.}

\item{plot}{Specifies whether a plot of the probability of testing recent over time should be produced}

\item{parallel}{Set to TRUE in order to perform bootstrapping in parallel on a multicore or multiprocessor syste. Not available on Windows.}

\item{cores}{Set number of cores for parallel processing when parallel=TRUE. This defaults to four.}
}
\value{
MDRI Dataframe containing MDRI point estimates, CI lower and upper bounds and standard deviation of point estimates produced during bootstrapping. One row per functional form.

Plots A plot of Probability of testing recent over time for each functional form.

Models The fitted GLM models for each functional form.
}
\description{
Estimate MDRI (point estimate and confidence interval) using binomial regression and a maximum likelihood approach
}
\details{
The package contains a long form vignette that covers the use of this function. Use vignette('TestCalibration',package='inctools') to access the vignette.

Expected data frame format: Before calling the function, please import your dataset into R environment.

time_var: Time since infection; Note: this package does not assume any specific time unit. It is important to specify the
recency time cut-off 'T' and the time-based data exclusion rule (inclusion_time_threshold) in the same unit as the input times.
The estimated MDRI will be in this unit.

Method: This function fits a function for probability of testing recent as a function of time to the supplied data using
binomial regression. This requires binary outcomes (recent/non-recent) coded as 1 for recent and 0 for non-recent test
resutls. Either a recency status variable must be specified, or a recency rule for determinging recency
status from a biomarker or set of biomarkers must be specified. Currently only independent biomarker
thresholds are supported (i.e. all biomarker criteria must be met in order for a specimen to be classified as recent).

Functional forms currently supported for the binomial regression fitting procedure:
cloglog_linear, logit_cubic

To be implemented in the near future: logit_spline

logit_cubic: Fits a binomial regression to probability of testing recent with a logit link on a polynomial in t of the
third degree, where t is time since (detectable) infection.

cloglog_linear: Fits a binomial regression to probability of testing recent with a log log link on log(t), where t is
time since (detectable) infection.

recency_rule: binary_data - supply a binary variable with 1=recent and 0=non-recent in recency_vars.

recency_rule:independent_thresholds: supply one threshold variable per biomarker in recency_vars and the relevant
thresholds, as well as whether a value below or above each threshold indicates recency in recency_params.

recency_params expects a list of pairs of thresholds and thresholdtypes, with zero indicating a reading below the threshold
 implies recency and 1 that a reading above the threshold implies recency. (Note: two values, a threshold and a
 thresholdtype per variable must be specified in recency_params. For example, if you specify recency_vars =
 c('ODn','ViralLoad') you may specify recency_params = c(1.5,0,500,1), meaning that an ODn reading below 1.5 AND a
 viral load reasing above 500 indicates a recent result. Objects with missing values in its biomarker readings will be
 excluded from caculation.
}
\examples{
mdri_ml_binomial(data=data, subid_var='PT_ID', time_var='DaysSinceEDDI',
 functional_forms=c('cloglog_linear','logit_cubic'), recency_rule='independent_thresholds',
 recency_vars=c('ODn','VL'), recency_params=c(1.5,0,400,1), n_bootstraps = 10000, alpha=0.05,
 plot=TRUE, parallel=FALSE)

In this example the MDRI and 95\% confidence interval would be calculated by determining
recency for each data point by evaluating whether normalised optical density (ODn) is
below 1.5 and Viral Load above 400 copies/ml. The probability of testing recent as a
function of time will be modelled using the loglog binomial (linear) and logit cubic polynomial
functional forms, the confidence interval will be estimated by means of 10,000
subject-level resampling operations and the fitted model curve will be plotted.
}

